==> DataProcessing/2_MovieCSVFilter.ipynb <==
import pandas as pd

ratings = pd.read_csv("MovieLensFilteredData/ratingsFiltered.csv")

ratings.shape

ratings.head(5)

# Create user csv

userDF = pd.DataFrame({"userId":ratings.userId.unique().tolist()})
userDF.head()

userDF.to_csv("MovieLensFilteredData/usersFiltered.csv", index=False)

# Create movie csv

eligibleMovies = ratings.movieId.unique().tolist()

movies = pd.read_csv('MovieLensRawData/movies.csv')
movies.head()

moviesFiltered = movies.loc[movies.movieId.isin(eligibleMovies)]
moviesFiltered.shape

# Merge with links

links = pd.read_csv('MovieLensRawData/links.csv')

links = links.loc[links['movieId'].isin(eligibleMovies)]

links.head(5)

links.loc[links['imdbId']==1285016]

agg = movies.merge(links, on='movieId', how='inner')
agg = agg.drop(columns=['tmdbId'])
agg.head()

agg.shape

agg.to_csv('MovieLensFilteredData/moviesFiltered.csv', index=False, header=True)


agg.loc[agg['imdbId']==118114]
