==> DataProcessing/5_ClusteringPrecomputed.ipynb <==
# Can we identify clusters in our movies

# import necessary libraries
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np

df = pd.read_pickle('Pickles/cos_sim_df.pkl')
# df = pd.read_pickle('Pickles/pca_100_df.pkl')


df.shape


# Dynamic Hierarchical Clustering
- we want our number of clusters to be dynamic
- we can do this by setting the distance threshold, such that only the movies above this threshold are valid for the cluster
- .1 would result in clusters being formed with >.9 cosine affinity

100/9904

df.head(2)

# Convert cosine similarity to distanc emetric

X = df.drop(columns=['movieId'], axis=1)

X = X.subtract(1)

X = X.abs()

X.head(2)

from sklearn.cluster import AgglomerativeClustering

# X = df.drop(columns=['Title'], axis=1)

# use distance threshold and n clusters is dynamic
clustering = AgglomerativeClustering(n_clusters=None, 
                                     affinity='precomputed', 
                                     linkage='complete', 
                                     compute_full_tree=True,
                                     distance_threshold=.5).fit(X)



cluster_labels = clustering.labels_.tolist() # store labels as a list
cluster_titles = df.movieId.tolist() # store titles as a list

# dict with our cluster number as key
cluster_dict = {}
for i,v in enumerate(cluster_labels):
    if v in cluster_dict:
        cluster_dict[v].append(cluster_titles[i])
    else:
        cluster_dict[v] = [cluster_titles[i]]
print("Number of clusters: {n}".format(n=len(cluster_dict)))

# dict with our movie title as key
title_dict = {}
for i,v in enumerate(cluster_titles):
    title_dict[v] = cluster_labels[i]


# Cluster Dataframe

- we need a dataframe with each cluster as a sample.

df_data_list = []
for k,v in cluster_dict.items():
    df_data_list.append({'cluster': k,
                         'movies': v})
    
cluster_df = pd.DataFrame(df_data_list)

cluster_df['cluster_size'] = [len(i) for i in cluster_df.movies]
cluster_df = cluster_df.sort_values(by='cluster_size', ascending=False)
cluster_df.head()

cluster_df.loc[cluster_df['cluster_size']==1].shape

cluster_df.cluster_size.mean()

# Pickle the cluster dataframe 

cluster_df.to_pickle('Pickles/clusterPrecomputed_df.pkl')
