==> DataProcessing/6_CreateUserItemMatrix.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('MovieLensFilteredData/ratingsFiltered.csv')

movieId_to_remove = 720

df = df.loc[df['movieId'] != movieId_to_remove]
df.shape

df.userId.unique()

df.to_csv('MovieLensFilteredData/ratingsFiltered.csv', index=False)

df.shape

# User

user_dict = {value: key for (key, value) in enumerate(df.userId.unique())}


df['userId_Index'] = df.apply(lambda x: user_dict[x['userId']], axis=1)


df.head()

# Movie

movie_dict = {value: key for (key, value) in enumerate(df.movieId.unique())}


df['movieId_Index'] = df.apply(lambda x: movie_dict[x['movieId']], axis=1)

df.head()

df.to_csv("MovieLensFilteredData/usermovieindex.csv", index=False)

# Create matrix

df = pd.read_csv("MovieLensFilteredData/usermovieindex.csv")

df = df.drop(columns=['userId', 'movieId', 'timestamp'])
df.head(1)

users = df.userId_Index.unique()
items = df.movieId_Index.unique()


# get number of users and items for shape of matrix
users_count = len(users)
items_count = len(items)
matrix_shape = (users_count, items_count)
matrix_shape

# create matrix of 0s
user_rating_matrix = np.zeros(shape=matrix_shape)
user_rating_matrix.shape

for i,v in df.iterrows():

    user_row = int(v.userId_Index)
    item_row = int(v.movieId_Index)
    rating = v.rating

    user_rating_matrix[user_row][item_row] = rating
    

user_rating_matrix.shape

np.save('Pickles/user_rating_matrix.npy', user_rating_matrix)    

user_rating_matrix
