==> DataProcessing/7_CreateClusterData.ipynb <==
import numpy as np
import pandas as pd
import pickle

# Load Data

### User Rating Matrix

user_rating_matrix = np.load('Pickles/user_rating_matrix.npy')  
user_rating_matrix.shape

### Cluster Dictionary

cluster_df = pd.read_pickle('Pickles/clusterPrecomputed_df.pkl')
cluster_df.shape

usermovieindex = pd.read_csv('MovieLensFilteredData/usermovieindex.csv')

original_indices = usermovieindex.movieId.unique()

new_indices = usermovieindex.movieId_Index.unique()

indices_dict = dict(zip(original_indices, new_indices))

def convert_index(row):
    output = []
    for movieId in row.movies:
        output.append(indices_dict[movieId])
    
    return output

cluster_df['new_indices'] = cluster_df.apply(lambda row: convert_index(row), axis=1)

cluster_df['cluster'] = cluster_df['cluster'].astype(str)


cluster_dict = dict(zip(cluster_df['cluster'], cluster_df['new_indices']))
len(cluster_dict)

with open('Pickles/cluster_dict.pickle', 'wb') as handle:
    pickle.dump(cluster_dict, handle, protocol=pickle.HIGHEST_PROTOCOL)
    
    
# with open('Pickles/cluster_dict.pickle', 'rb') as handle:
#     b = pickle.load(handle)
    
# print(len(b))

### Similarity Matrix

sim_matrix = pd.read_pickle('Pickles/cos_sim_df.pkl')
sim_matrix.shape

# Fit algorithm
